+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add `RBlade.direct_component_rendering` option to allow RBlade components to be rendered directly
 - Add `component` view helper for rendering RBlade components in other templates
 - Improve performance of regular expressions
+- Remove deprecated "_required" option for props statement values (use "required" instead)
 - Update README
 - Fix issue with component tokenizer hanging